+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/take/AbstractTakeAction.java
@@ -2,12 +2,14 @@ package org.betonquest.betonquest.quest.action.take;
 
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.QuestException;
+import org.betonquest.betonquest.api.instruction.Argument;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.action.OnlineAction;
 import org.betonquest.betonquest.database.PlayerData;
 import org.betonquest.betonquest.quest.action.NotificationSender;
 import org.bukkit.entity.Player;
+import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 
 import java.util.ArrayList;
@@ -24,7 +26,7 @@ public abstract class AbstractTakeAction implements OnlineAction {
     /**
      * The order in which the checks should be performed to remove the item.
      */
-    protected final List<CheckType> checkOrder;
+    protected final Argument<List<CheckType>> checkOrder;
 
     /**
      * The notification sender to send notifications to the player.
@@ -37,7 +39,7 @@ public abstract class AbstractTakeAction implements OnlineAction {
      * @param checkOrder         the order in which the checks should be performed
      * @param notificationSender the notification sender to use
      */
-    public AbstractTakeAction(final List<CheckType> checkOrder, final NotificationSender notificationSender) {
+    public AbstractTakeAction(final Argument<List<CheckType>> checkOrder, final NotificationSender notificationSender) {
         this.checkOrder = checkOrder;
         this.notificationSender = notificationSender;
     }
@@ -49,11 +51,12 @@ public abstract class AbstractTakeAction implements OnlineAction {
      * @throws QuestException if an error occurs during the check
      */
     protected void checkSelectedTypes(final OnlineProfile profile) throws QuestException {
-        for (final CheckType type : checkOrder) {
+        for (final CheckType type : checkOrder.getValue(profile)) {
             switch (type) {
                 case INVENTORY -> checkInventory(profile);
                 case ARMOR -> checkArmor(profile);
-                case OFFHAND -> checkOffhand(profile);
+                case MAINHAND -> checkEquipmentSlot(profile, EquipmentSlot.HAND);
+                case OFFHAND -> checkEquipmentSlot(profile, EquipmentSlot.OFF_HAND);
                 case BACKPACK -> checkBackpack(profile);
             }
         }
@@ -73,14 +76,14 @@ public abstract class AbstractTakeAction implements OnlineAction {
         player.getInventory().setArmorContents(newArmor);
     }
 
-    private void checkOffhand(final OnlineProfile profile) {
+    private void checkEquipmentSlot(final OnlineProfile profile, final EquipmentSlot slot) {
         final Player player = profile.getPlayer();
-        final ItemStack offhand = player.getInventory().getItemInOffHand();
-        final ItemStack[] newOffhand = takeDesiredAmount(profile, offhand);
-        player.getInventory().setItemInOffHand(newOffhand[0]);
+        final ItemStack item = player.getInventory().getItem(slot);
+        final ItemStack[] newItem = takeDesiredAmount(profile, item);
+        player.getInventory().setItem(slot, newItem[0]);
     }
 
-    private void checkBackpack(final OnlineProfile profile) throws QuestException {
+    private void checkBackpack(final OnlineProfile profile) {
         final PlayerData playerData = BetonQuest.getInstance().getPlayerDataStorage().get(profile);
         final List<ItemStack> backpack = playerData.getBackpack();
         final List<ItemStack> newBackpack = removeDesiredAmount(profile, backpack);