+++ CHANGELOG.md
@@ -65,6 +65,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `craftEngine` item type
 - `itemsAdder` item type
 - `brew` item type
+- `mainhand` selection order to `take` action
 ### Changed
 - Spigot is no longer supported, paper is now required 
 - message.yml file was deleted and instead the lang folder now contains all translations